+++ ConnectorAutocadCivil/AutocadCivilDUI3/AutocadCivilDUI3Shared/AutocadCivilDUI3Shared.projitems
@@ -19,11 +19,11 @@
       <DependentUpon>DUI3PanelWebView.xaml</DependentUpon>
     </Compile>
     <Compile Include="$(MSBuildThisFileDirectory)Entry.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Extensions\Editor.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Extensions\Extends3d.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Utils\ApplicationIdManager.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Utils\AutocadDocumentManager.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Utils\AutocadDocumentModelStore.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Utils\EditorExtensions.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Utils\Extends3dExtensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Utils\Objects.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Utils\TransactionContext.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Utils\Utils.cs" />
