+++ ConnectorRhino/ConnectorRhinoWebUI/Bindings/BasicConnectorBinding.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using ConnectorRhinoWebUI.Extensions;
 using ConnectorRhinoWebUI.Utils;
 using DUI3;
 using DUI3.Bindings;
@@ -58,13 +59,13 @@ public class BasicConnectorBinding : IBasicConnectorBinding
 
   public void UpdateModel(ModelCard model)
   {
-     var idx = _store.Models.FindIndex(m => model.Id == m.Id);
+     int idx = _store.Models.FindIndex(m => model.Id == m.Id);
     _store.Models[idx] = model;
   }
   
   public void RemoveModel(ModelCard model)
   {
-    var index = _store.Models.FindIndex(m => m.Id == model.Id);
+    int index = _store.Models.FindIndex(m => m.Id == model.Id);
     _store.Models.RemoveAt(index);
   }
 
@@ -73,26 +74,12 @@ public class BasicConnectorBinding : IBasicConnectorBinding
     SenderModelCard model = _store.GetModelById(modelCardId) as SenderModelCard;
     List<string> objectsIds = model.SendFilter.GetObjectIds();
     List<RhinoObject> rhinoObjects = objectsIds.Select((id) => RhinoDoc.ActiveDoc.Objects.FindId(new Guid(id))).ToList();
-
+    
     RhinoDoc.ActiveDoc.Objects.UnselectAll();
-    foreach (var rhinoObject in rhinoObjects)
-    {
-      RhinoDoc.ActiveDoc.Objects.Select(rhinoObject.Id);
-    }
+    RhinoDoc.ActiveDoc.Objects.Select(rhinoObjects.Select(o => o.Id));
     
     // Calculate the bounding box of the selected objects
-    BoundingBox boundingBox = BoundingBox.Unset;
-    foreach (var obj in rhinoObjects)
-    {
-      BoundingBox objBoundingBox = obj.Geometry.GetBoundingBox(false);
-      if (objBoundingBox.IsValid)
-      {
-        if (boundingBox.IsValid)
-          boundingBox.Union(objBoundingBox);
-        else
-          boundingBox = objBoundingBox;
-      }
-    }
+    BoundingBox boundingBox = BoundingBoxExtensions.UnionRhinoObjects(rhinoObjects);
     
     // Zoom to the calculated bounding box
     if (boundingBox.IsValid)
