+++ ConnectorRhino/ConnectorRhinoWebUI/ConnectorRhinoWebUI.csproj
@@ -45,6 +45,7 @@
     <Compile Remove="SpeckleWebUiPanelWebView2.xaml.cs" />
   </ItemGroup>
   
+  
   <PropertyGroup Condition="$(Configuration) == 'Debug' AND $([MSBuild]::IsOSPlatform(Windows))">
     <StartProgram>C:\Program Files\Rhino 7\System\Rhino.exe</StartProgram>
     <StartArguments></StartArguments>
